+++ app/src/main/java/pt/ipleiria/estg/tesp/psi/projsi/sistematrocas/projetosi_android_1718/modelos/Eletronica.java
@@ -8,7 +8,7 @@ public class Eletronica extends Categoria {
     private String descricao;
     private String marca;
 
-    public Eletronica(Integer id, String nome, String descricao, String marca) {
+    public Eletronica(Long id, String nome, String descricao, String marca) {
         super(id, nome);
         this.descricao = descricao;
         this.marca = marca;
