+++ app/src/main/java/pt/ipleiria/estg/tesp/psi/projsi/sistematrocas/projetosi_android_1718/modelos/Roupa.java
@@ -5,11 +5,13 @@ package pt.ipleiria.estg.tesp.psi.projsi.sistematrocas.projetosi_android_1718.mo
  */
 
 public class Roupa extends Categoria {
+
     private String marca;
     private String tamanho;
-    private Integer idTipo;
+    private Long idTipo;
 
-    public Roupa(Integer id, String nome, String marca, String tamanho, Integer idTipo) {
+    public Roupa(Long id, String nome, String marca, String tamanho, Long idTipo) {
+
         super(id, nome);
         this.marca = marca;
         this.tamanho = tamanho;
@@ -32,11 +34,11 @@ public class Roupa extends Categoria {
         this.tamanho = tamanho;
     }
 
-    public Integer getIdTipo() {
+    public Long getIdTipo() {
         return idTipo;
     }
 
-    public void setIdTipo(Integer idTipo) {
+    public void setIdTipo(Long idTipo) {
         this.idTipo = idTipo;
     }
 }
